SpringBoot: JPA Buddy

+++ SpringBoot/store/src/main/resources/db/migration/V6__add_categories_and_products.sql
@@ -1,6 +1,6 @@
 CREATE TABLE categories
 (
-    id   TINYINT AUTO_INCREMENT PRIMARY KEY,
+    id   BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL
 );
 
@@ -9,7 +9,7 @@ CREATE TABLE products
     id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     name        VARCHAR(255)   NOT NULL,
     price       DECIMAL(10, 2) NOT NULL,
-    category_id TINYINT,
+    category_id BIGINT,
     CONSTRAINT fk_category
         FOREIGN KEY (category_id) REFERENCES categories (id)
             ON DELETE RESTRICT
